Refactor settingsStore into dedicated file
